uccp420wlan: Fix old production mode
With the recent changes to get_rate() function the old
production mode is broken.

The variable all_rates_invalid is not reset hence the
default rates are taken instead of taking from proc, but
rate_flags are taken from proc this causes an RPU lockup.

a) Reorder the code to handle this.
b) reset rate_flags in defaults
c) use a different rate_retries to differentiate the
default code (empty/invalid rates)

Change-Id: I3e33959ee39a73077c30a3bbfd3fffae32ecf4fc